ShippingMethodDataX to ShippingDataX
