+++ assets/js/base/context/cart-checkout/shipping/constants.js
@@ -1,7 +1,7 @@
 /**
  * @typedef {import('@woocommerce/type-defs/contexts').ShippingErrorTypes} ShippingErrorTypes
  * @typedef {import('@woocommerce/type-defs/cart').CartShippingAddress} CartShippingAddress
- * @typedef {import('@woocommerce/type-defs/contexts').ShippingMethodDataContext} ShippingMethodDataContext
+ * @typedef {import('@woocommerce/type-defs/contexts').ShippingDataContext} ShippingDataContext
  */
 
 /**
@@ -29,7 +29,7 @@ export const DEFAULT_SHIPPING_ADDRESS = {
 };
 
 /**
- * @type {ShippingMethodDataContext}
+ * @type {ShippingDataContext}
  */
 export const DEFAULT_SHIPPING_CONTEXT_DATA = {
 	shippingErrorStatus: ERROR_TYPES.NONE,
